Restructure code, refactor unit test-cases and incorporate review comments
